Backup commit before big changes
I've noticed that agents created in different states cannot communicate with each other and instead of complicating the solution to make it work I've decided to change the entire project and stop using agents. If for some reason I decide to return to this implementation I'd have to create / implement transition state in which a manager can act as a router -> receive agent messages analyse them and forward them to agent that should receive them. (This is unnecessary complication in my opinion)