Added scan analysis
The script now uses expoitdb and searchsploit to find a list of possible exploits for nmap scan results. Currently the json produced by searchsploit can't really be loaded and a manual parser has to be written for it. Parser should open the file, remove [i] lines and convert the text to correct json format.  Added support for nmap result analysis to reportgenerator (just a method, not implemented yet). Changed scan options and which results are returned. Scan now creates xml output as well (not just csv). The nmapanalyser.py uses the nmap xml report that was generated to search for potential exploits. Added ascii art on script startup, with random font on each run.