Added msfrpc implementation
Updated the dockerfile, fixed a bug in pentest.py where instead of passing nmap's xml report filename full csv report was passed as an argument to the nmapanalyser. More importantly, I've added most of the backbone of the msf rpc implementation in metasploitmanager.py. I've still got to implement means of connecting to the database in the msf client session. What is left to do is to connect to an existing db / wipe the db and then test whether the db_import will work