Added console wrappers
Added a pymetasploit wrapper method for sending commands to the msf and waiting for it's output. (The pymetasploit3 has something similar implemented with run_module_with_output but in this case this one is more suitable for this usage).  Added two main groups of TODO's / ideas how to proceed with the selection of the proper auxiliary scanner and it's usage. The method used for this will probably be the same for the exploit stage as well. (First method is to define a number N of common protocols which for which services comand output will be checked and specific auxiliary modules will be used. Second option is to try and run a scan for each available auxiliary scanner module for that service. The first approach is faster but second is more thorough). Added a list of tools for which a support would be great (they're supported in msf and as such they could probably be used as well e.g. nessus).