Fixed module selection
The main functionality of this program is now almost done. Metasploitmanager.py can now run metasploit's auxiliary scanner or agent modules if the service detected can be analysed with any of them. Still have to test if it is going to work with exploit modules. Current problem is that there are too many false positives (for example, if a service is described as ftp, the program will try to scan a target with every ftp scanner on the list that can be run in the given context). Added required methods to find modules, check if they can be used and methods to run them. Added more information about the process (the program is a bit more verbose now). Refactored the pentest.py so that it doesn't output nmap scan report anymore, and all inputs and outputs are tool report's filenames.