Lowered aux modules count
Changed how the aux modules were being selected. Now the problem is that not enough effective modules are being run (most of them return errors but at least the report file isn't huge) and as such I've decided that I'm going to use this selection process for exploitation process, and that there's going to be a separate class dedicated to auxiliary scanning for some of the more common protocols with modules that actually give some interesting reports (based on detected services). Current methods should probably work with exploit modules of msf.