Refactor and upgrade
Refactored the code and created classes for running of auxiliary scans in metasploit, class for automated execution of exploits, class for communicating with msfconsole and class for outputting msf activity reports. Renamed scanner.py into nmapscanner to remove confusion about it's responsibility. The automatic exploitation part still needs a bit more work so that it doesn't crash.