Code refinement
Added a method to create reports folder so that the reports can be found in the same space.  Added initialization and tear-down methods which aren't tested and aren't used at the moment. The problem is that sub-processes are being started but the code doesn't wait for started programs to initialize and thus the program crashes because msfrpcd hasn't been initialized.  Improved the program's verbosity by adding more informations about the whereabouts of generated reports.  Created session info functionality added so that the info is provided when running exploits. Added created session auto-kill so that the next exploit could start the exploitation process.